tests/requirements.txt
pytest==7.1.2
requests==2.28.1
fastapi==0.86.0
pydantic==1.10.2
python-jose==3.3.0